Update getDiscretization for newer OpenSees
The previous method for getting the fiber information doesn't work in
newer versions of OpenSees. The same information can be obtained by
using the JSON output commands, but not all materials support JSON
output.

A new property fiber_discretization_method is added, with two possible
values: 'legacy' and 'json' (defaults to 'json'). This way the legacy
code is still available for now.